+++ Sources/names.h
@@ -1,10 +1,11 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define NO_NAME 0
 
-typedef uint64_t name_id;
+typedef size_t name_id;
 
 void names_init(void);
